Initial commit.
Setup the project structure as well as Travis CI integration.  This commit
also introduces support for building any dropwizard-bundles project for a
specific version of Dropwizard. It also ensures that each bundle is coded
in a consistent way and that dependencies that conflict with Dropwizard
aren't accidentally introduced into a bundle.